requirements.txt
python-telegram-bot>=20.3
yt-dlp>=2024.3.10
md2tgmd>=0.1.2
Pillow>=10.3.0
